+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -1,6 +1,7 @@
 package lotto.domain;
 
 import lotto.domain.number.LottoNumbers;
+import lotto.domain.price.Price;
 import lotto.domain.prize.LottoPrizeResult;
 import lotto.domain.prize.Prize;
 import lotto.domain.ticket.LottoTicket;
