+++ src/test/java/lotto/domain/price/PriceTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.price;
 
 import lotto.exception.NegativeNumberException;
 import org.junit.jupiter.api.DisplayName;
